+++ lib/dockerfile_distro.mli
@@ -23,10 +23,10 @@
 (** {2 Known distributions and OCaml variants} *)
 
 type t = [ 
-  | `Alpine of [ `V3_3 ]
+  | `Alpine of [ `V3_3 | `V3_4 ]
   | `CentOS of [ `V6 | `V7 ]
   | `Debian of [ `V9 | `V8 | `V7 | `Stable | `Testing | `Unstable ]
-  | `Raspbian of [ `V8 | `V7 ]
+  | `Raspbian of [ `V8 ]
   | `Fedora of [ `V21 | `V22 | `V23 ]
   | `OracleLinux of [ `V7 ]
   | `Ubuntu of [ `V12_04 | `V14_04 | `V15_04 | `V15_10 | `V16_04 | `V16_10 ]